Bug 1261715 - Avoid promise rejections in browser_rules_add-rule_05.js; r=jryans
This test simulates a context click in the rule view and selects the
"add rule" item in order to add a new rule.
Adding a rule triggers a ruleview-changed event, and so this test should
start listening for this event before triggering the action, which it
did not do.

MozReview-Commit-ID: BPO8Fgwsq5M

UltraBlame original commit: 8046e825980ac4080619527b72292b8e9ac1f56d