Bug 1213670 - Remove dom/imptests/ that aren't likely to be platform-specific; r=jgraham
These are all duplicated in testing/web-platform/, so the only reason to
keep them is that web-platform-tests don't yet run on every platform
that mochitests do.  But they get in the way, so let's remove the ones
that are relatively unlikely to be platform-specific.

UltraBlame original commit: 349594185facf0e5e093be7c71b3be8e621db55d