Bug 1265038 - Turn on Adobe EME by build config rather than per-platform basis. r=jwwang
We end up showing the "Adobe EME is preffed off" UI on MacOSX because we're
check for Adobe EME support based on platform, not based on whether it's
configured on in the mozconfig. We should really only check for Adobe EME
if 'ac_add_options --enable-eme=adobe' is present in the build config,
as otherwise preffing it on isn't going to work.


MozReview-Commit-ID: CpSAh6gUU1S

UltraBlame original commit: 5bf181aae1ad9aba0ec64a4e0b6d5ab23476562c