Bug 1243538: [webm] P7. Let the decoder handle picture resizing. r=SingingTree
Is is now up to the decoders to automatically determine the correct size of the decoded frame.
By leaving the original dimensions in the VideoInfo, it will allow automatic rescale of the cropping window according to the original dimensions.

MozReview-Commit-ID: FZM4YlataNz

UltraBlame original commit: 1687a256328b8f3b3476b1fa5053e1f946134ca7