Bug 1053986 - Rename nsIFrame::IsHorizontal to IsXULHorizontal, and related methods.  r=dholbert
This is a manual subset of changes written with sed, over .h and .cpp
files in layout/.

This also renames a static method on nsSprocketLayout.

Note that nsFlexContainerFrame and nsRangeFrame also have IsHorizontal
methods that are not renamed here, but this can be found to be
relatively safe because none of the IsHorizontal methods are virtual.

MozReview-Commit-ID: Jsdy7I4Q7mX

UltraBlame original commit: a626ff70c178b0627d9693301b5438bd5e81dda5