Bug 1267930 - When the wasapi rendering loop is stuck and we're shuttin down, leak the thread and continue the shutdown process. r=kinetik
It's probably better to not TerminateThread here, because we don't really know
what happened down the stack of the rendering loop, so we just leak it. This
happens during shutdown so it's not really an issue, it's going to be cleaned
up by the OS anyways.

MozReview-Commit-ID: 4o3Krj7L7m2

UltraBlame original commit: 8a9107830c844b4b8f811463ca6f9c30802d5acf