Bug 1265433 - Move bookmarks star into top bar of menu r=sebastian
We do this by showing history items in the "share" menu item, and
removing the existing "quickshare" item that showed history items.

We also need to shuffle our icon sizes around: our xhdpi icon size matches
the hdpi icon sizes of the other icons in the top bar (modulo padding, which
we need to add as part of this commit), similarly our xxhdpi icon is shifted to
xhdpi. We don't supply an xxhdpi icon at all for any of the icons in the top bar
of the menu, hence that size is removed (and reused for the xhdpi icon, see above).

MozReview-Commit-ID: 7fabPQ5Tyst

UltraBlame original commit: 8faf7ba00b3148a01db1701e80c1c3e611c4066d