Bug 1266022 - skip checking of LoadInfo for chrome resources. r=sicking
When tab with a userContextId!=0 contains a chrome page (such as
about:newtab) the userContextId in the LoadInfo won't match the
userContextId in the LoadContext. The LoadInfo will contain the
systemPrincipal and so use userContextId=0, the LoadContext has the
userContextId of the tab (!=0).

This is fine as long as we page only loads chrome-URLs and other
non-http URLs since those don't use cookies anyway.

So avoid asserting in this situation.

Long term we want the chrome page to use the default userContextId for
cookies, since that's what it chrome code normally use. This will work
properly once we get the cookie jar information from the LoadInfo rather
than from the LoadContext.

UltraBlame original commit: 5fbcba6455cb98a95705249606afbe6fb53cdc3d