Bug 1269774 - Reduce browser menu top padding r=sebastian
We want to have equal paddings for the top and right hand side of the
menu, whereas we previously had 24dp padding for the top, and only 8dp
on the right.

Note: the padding is defined in the 9-patch files, this commit simply
removes empty vertical space from the top of the menu background images.

MozReview-Commit-ID: Db0hXsPOVpp

UltraBlame original commit: 4067fdc3fdc04df8033ee532caff07b995fef18a